Fix issue #41: Refactor models file
