+++ photo/models/soft_delete_model.py
@@ -1,0 +1,20 @@
+from django.db import models, transaction
+from photo.manager import SoftDeleteManager
+
+
+class SoftDeleteModel(models.Model):
+    is_deleted = models.BooleanField(default=False)
+    objects = SoftDeleteManager()
+    all_objects = models.Manager()
+
+    @transaction.atomic
+    def delete(self):
+        self.is_deleted = True
+        self.save()
+
+    def restore(self):
+        self.is_deleted = False
+        self.save()
+
+    class Meta:
+        abstract = True
